requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
filelock==3.6.0
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
numpy==1.22.3
packaging==21.3
pyparsing==3.0.8
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.10.0+cu113
tqdm==4.64.0
.utils/transformers-4.16.2
typing_extensions==4.1.1
urllib3==1.26.9